modular programming component
